add new parameter server client module
