+++ nd4j-parameter-server-parent/nd4j-parameter-server-client/src/test/resources/log4j.properties
@@ -1,0 +1,23 @@
+
+log4j.rootLogger=ERROR, Console
+log4j.logger.play=DEBUG
+log4j.appender.Console=org.apache.log4j.ConsoleAppender
+log4j.appender.Console.layout=org.apache.log4j.PatternLayout
+log4j.appender.Console.layout.ConversionPattern=%d{ABSOLUTE} %-5p ~ %m%n
+
+log4j.appender.org.springframework=DEBUG
+log4j.appender.org.nd4j=INFO
+log4j.logger.org.nd4j.aeron.ipc=INFO
+log4j.appender.org.canova=INFO
+log4j.appender.org.deeplearning4j=INFO
+log4j.appender.opennlp.uima=OFF
+log4j.appender.org.apache.uima=OFF
+log4j.appender.org.cleartk=OFF
+
+log4j.logger.org.springframework=INFO
+log4j.logger.org.nd4j=DEBUG
+log4j.logger.org.deeplearning4j=INFO
+log4j.logger.opennlp.uima.util=OFF
+log4j.logger.org.apache.uima=OFF
+log4j.logger.org.cleartk=OFF
+
